+++ openfisca_ceq/tools/indirect_taxation/__init__.py
@@ -1,0 +1,1 @@
+# -*- coding: utf-8 -*-
