+++ openfisca_ceq/tools/indirect_taxation/variables_generator.py
